requirements.txt
django
djangorestframework
djangorestframework-simplejwt
dj-database-url
whitenoise
gunicorn